use size_t in audio.c
Buffer sizes should be size_t.  This is safe here, at least not
unsafer than without the patch.  I have no idea why audioBufferSize
and audioBufferPos were explicitly declared as signed integer.

git-svn-id: https://svn.musicpd.org/mpd/trunk@7296 09075e82-0dd4-0310-85a5-a0d7c8717e4f

+++ src/audio.h
@@ -59,7 +59,7 @@ void finishAudioDriver(void);
 
 int openAudioDevice(AudioFormat * audioFormat);
 
-int playAudio(const char *playChunk, int size);
+int playAudio(const char *playChunk, size_t size);
 
 void dropBufferedAudio(void);
